[#1626](feat) Add overview rectangle in main view
