+++ application/client/src/app/ui/views/toolbar/chart/chart/state.ts
@@ -5,6 +5,7 @@ import {
     IPosition,
     IPositionChange,
     UPDATE_TIMEOUT_MS,
+    IRectangle,
 } from '../common/types';
 import { IRange } from '@platform/types/range';
 import { Owner } from '@schema/content/row';
@@ -40,6 +41,14 @@ export class State extends AbstractState {
         width: this._canvasWidth,
     };
     private _zoomedRange!: IRange;
+    private readonly _rectangle: IRectangle = {
+        width: 0,
+        left: 0,
+    };
+    private _shownRange: IRange = {
+        from: 0,
+        to: 0,
+    };
 
     constructor() {
         super();
@@ -63,6 +72,7 @@ export class State extends AbstractState {
         );
         this._initializeSubscriptions();
         this._onResize();
+        this._updateOverviewRectangle(this._session.cursor.frame().get());
     }
 
     public destroy() {
@@ -76,6 +86,10 @@ export class State extends AbstractState {
 
     public get hasNoData(): boolean {
         return this._dataState.filter.hasNoData && this._dataState.chart.hasNoData;
+    }
+
+    public get rectangle(): IRectangle {
+        return this._rectangle;
     }
 
     public onClick(event: MouseEvent) {
@@ -133,6 +147,9 @@ export class State extends AbstractState {
                 this._parent.ilc().channel.ui.sidebar.resize(this._onResize.bind(this)),
                 this._parent.ilc().channel.ui.window.resize(this._onResize.bind(this)),
                 this._service.subjects.change.subscribe(this._onZoomChange.bind(this)),
+                this._session.cursor.subjects
+                    .get()
+                    .frame.subscribe(this._updateOverviewRectangle.bind(this)),
             );
     }
 
@@ -167,7 +184,27 @@ export class State extends AbstractState {
                 width: this._canvasWidth,
             };
             this._filter.zoom(this._zoomedRange);
+            this._updateOverviewRectangle();
             this._timeout.canvasWidth = -1;
         }, 150);
+    }
+
+    private _updateOverviewRectangle(range?: IRange) {
+        if (range !== undefined) {
+            range.to += 1;
+            this._shownRange = range;
+        }
+        const linesDiff: number = this._shownRange.to - this._shownRange.from;
+        const linesTotal: number = this._session.stream.len();
+        const linesDiffPercent: number = linesDiff / linesTotal;
+        const fromPercent: number = this._shownRange.from / linesTotal;
+        const border: number = 2;
+
+        this._rectangle.width = Math.round(linesDiffPercent * this._canvasWidth);
+        this._rectangle.left = Math.round(fromPercent * this._canvasWidth);
+        if (this._rectangle.left + this._rectangle.width + border > this._canvasWidth) {
+            this._rectangle.left = this._canvasWidth - this._rectangle.width - border;
+        }
+        this._parent.detectChanges();
     }
 }
