+++ application/client/src/app/ui/views/toolbar/chart/chart/styles.less
@@ -33,4 +33,11 @@
     & .hidden {
         visibility: hidden;
     }
+    & div.rectangle {
+        height: calc(100% - 2px);
+        bottom: 0px;
+        pointer-events: none;
+        position: absolute;
+        border: dashed @scheme-color-1 thin;
+    }
 }
