+++ application/client/src/app/ui/views/toolbar/chart/styles.less
@@ -20,7 +20,7 @@
     }
     & div.divider {
         border: dashed @scheme-color-1 thin;
-        bottom: calc(3rem + 5px);
+        bottom: calc(3rem + 7px);
         position: absolute;
         width: 100%;
     }
@@ -31,11 +31,5 @@
         right: 0;
         bottom: 0;
         overflow: hidden;
-    }
-    & div.rectangle {
-        height: 99%;
-        pointer-events: none;
-        position: absolute;
-        border: dashed @scheme-color-1 thin;
     }
 }
