+++ application/client/src/app/ui/views/toolbar/chart/template.html
@@ -5,4 +5,3 @@
 <div class="secondary">
     <app-views-chart-zoomer-canvas [session]="session"></app-views-chart-zoomer-canvas>
 </div>
-<div *ngIf="!state.hasNoData" class="rectangle" [ngStyle]="{'width': state.rectangle.width + 'px', 'left': state.rectangle.left + 'px'}"></div>
