+++ application/client/src/app/ui/views/toolbar/chart/zoomer/styles.less
@@ -26,4 +26,11 @@
         opacity: 0;
         pointer-events: none;
     }
+    & div.rectangle {
+        height: calc(100% - 2px);
+        bottom: 0px;
+        pointer-events: none;
+        position: absolute;
+        border: dashed @scheme-color-1 thin;
+    }
 }
